Fixed the float-to-int bug, but the bots are still inactive when their candle is out. FredrikBoii has already made a fix for a similar issue which also solves this bug. Pushing to save progress.

+++ Sail/src/Sail/entities/components/CandleComponent.cpp
@@ -9,7 +9,7 @@ CandleComponent::~CandleComponent() {
 
 }
 
-void CandleComponent::hitWithWater(int damage) {
+void CandleComponent::hitWithWater(float damage) {
 	m_damageTakenLastHit = damage;
 	m_wasHitByWater = true;
 }
